Build Semaphore: Working prototype.
Proof of concept. DO NOT MERGE!

Create a semaphore, to restrict the number of builders that may enter
the "build" phase of their ports. This number is controlled by the
"BUILD_SEMAPHORE" setting. With a setting of e.g. "BUILD_SEMAPHORE=3",
no more than 3 builders are allowed to enter the "build" phase at a
time. Other phases are unaffected, they are only limited by the total
number of builders ("PARALLEL_JOBS").