image_zfs: Allow same-as-host pool names
Use `zpool create -t` to set the in-core pool name to a temporary,
random pool name, while the on-disk name will be the name specified as
the pool name in `ZFS_POOL_NAME`, by default `zroot` to match what is
currently used on release images.

Keep the check for a possible name collision on the randomly-created
pool name, to err on the side of caution.

Inspired by:	https://reviews.freebsd.org/D34426