pkgclean: Add -C to allow deleting only specific packages.
Also add some clarifications in the manpage and output about
how listed packages will be interpreted for deletion.

The -C mode will only delete specified packages - it will not
touch their dependencies even if they are orphaned. A second
run without -C is needed to remove orphaned dependencies. This
is because -C is intended for adhoc deletions where the package
is still *wanted* but needs to be rebuilt.

Fixes #310