Remove obsolete test for vfs.mnt_free_list_batch
In 2016 batching was introduced for free list management. Then in a59937e
poudriere started checking of the sysctl is present to manage parallel unmounts.

However, as of 2020 starting with:

  commit cc3593fbd953e2eeec72dbca05e77632817cba0b
  Author: Mateusz Guzik mjg@FreeBSD.org
  Date: Mon Jan 13 02:37:25 2020 +0000

  vfs: rework vnode list management

the sysctl is no longer present and the test no longer makes sense -- the
oldest supported release already has the sysctl and newer ones don't
while not having the perf problem.

Fixes #1011