Drop DEPENDS_ARGS support.
This was never actually used. It was a pre-FLAVORS thing that attempted
to support multiple python versions. Support for it is explicitly
disabled if the ports tree has FLAVORS support. No one has asked about
supporting this along with FLAVORS. FLAVORS is many years old now.
The feature incurs a significant complexity into the metadata lookup and
queue.
Time to drop it.