clean_pool: Remove some dead IGNORED handling.
This was added in 5d6aef4b8f but is not needed due to both
gather_ports_vars() and compute_deps() not processing deps of IGNORED ports.
Handling for those came in da49bc02a8 and 06b67be81c01b9. Primarily
it is due to compute_deps() not relating the IGNORED dependency to
the port that would have been skipped if it were not for it already
being IGNORED.

Add some extra comments about this.

Tweak the bulk-ignore-and-skipped.sh test for this slightly to capture
an ordering requirement, and add comments explaining the expectations.