Work around sh(1) bug with interrupt during redirecting to a pipe.
An sh(1) bug exists where an EINTR during a redirect (for example, writing to a
pipe with no reader) will jump straight to the EXIT trap. This also unsets the
EXIT trap to avoid recursion. After it executes the first line it detects the
interrupt and jumps to that handler. It never returns to the EXIT trap.
Resetting the handler here will ensure we return back to it.