image_zfs: Fix GPT+ZFS boot partition creation
Apparently #962 never really fixed the issue, just avoided a failed
assertion.

Fix it by making a disposable copy of the gptzfsboot file, and truncate
(or rather expand) it to 512K before passing it to mkimg as the
freebsd-boot partition, in order to keep the same partition size that
bsdinstall uses.

Fixes:	#1098