Rework how/when stats_queued is calculated.
This reworks what was done in commit 77cbefc72501df for Issue #727.
The count is restored to before that change but now it should not
be possible to see negatives.

The .poudriere.ports.queued and stats_queued contains everything that
will be built and everything that was *requested* to be built, even if
those don't need to be built or are IGNORED. The purpose of it is for
the HTML interface to see what the beginning of the job looked like
before the hard work began.

This also fixes tests to consider IGNORED ports as part of the queue.