Add some :? to vars that may accidentally touch the host
A lot of this is redundant or not needed as there are not-empty checks
for the specific cases. However until we use set -u we should adopt this
style to avoid any mistakes from typos or refactoring.

Issue #977

In some cases ? is used as the command may be ran in the host.