logclean: Rework locking for increased testing parallelism.
This also introduces a lock into the build on the logs they are writing
to prevent logclean from touching them.

Lower HTML json lock wait time.

Add -w flag to specify how long to wait for the global logclean lock,
defaulting to not waiting as normal usage won't have this lock held.

This is after finding this on my system:
  Filesystem    Size    Used   Avail Capacity  Mounted on
  tmpfs         126G    124G    2.3G    98%    /tmp