Rework queue to have both build and run jobs.
This allows the job queue to have smarter dependencies for ordering
jobs. This will allow things like TEST_DEPENDS, and fixes some issues
with TRIM_ORPHANED_BUILD deps (which the removed XFAIL tests test for).

Some job types do not do anything except provide ordering. They are
trimmed from the queue at the time of running.

Rework the queue to support just-in-time determination of whether a port needs
to build or not.  That is, on startup do not remove items from the queue that
are not needed.  They will be trimmed out when eligible to be ran.  This incurs
slight overhead but avoids a lot of problems with determining the proper
dependency graph needs or optimizing it early on which can do the wrong thing.

This is primarily needed to support PKG_NO_VERSION_FOR_DEPS which avoids
deleting packages that do not need a rebuild (effectively defaults
SKIP_RECURSIVE_REBUILD to 1.  In that case we ended up in a situation where
some dependencies in the middle of the graph did not need a rebuild but items
further down the chain did.  The trimming would incorrectly clear out the
middle down, losing the needed rebuild.  An example is someport -> patchutils
-> bash.  If patchutils did not need a rebuild, but bash did, the queue was
trimming out both patchutils and bash.  Determining what needed to be kept
pre-build was overly complex.  With this method we determine the need on visit
right before build.

Fixes #806 - The TRIM_ORPHANED_BUILD_DEPS problem and similar solution.
Issue #822