poudriere.css: Adapt to latest DataTables style to avoid unreadable tables
The newer DataTables imported in ce2146e99efc ("Update DataTables to
1.13.8") started setting every other row to transparent. Since we add
the built etc classes to the whole table, this means the entire table's
background colour is green etc, and so the transparent rows now allow
that colour to shine through, making it hard to read the text.

Adapt the CSS selectors to only colour the thead within the whole table
so that the rows are restored to being uncoloured, less garish and
rather more readable. Note that these classes are also used for the
status display at the top, so we need to handle them on things other
than tables too, where we continue to just colour the whole element.

Fixes #1114

Fixes:	ce2146e99efc ("Update DataTables to 1.13.8")