runtest: Fix gmake jobserver race.
This fixes a race when using the runtest's own job token
for running the child test. The race could cause it to
wait forever on the job server pipe for a new token
when it will never receive one.

This could be seen with something like -j3 and running
calculate_duration.sh and read_file.sh tests.