pkgbase: Fix building packages
When updating a jail we first make buildworld/buildkernel and then
we make update-packages.
Since the src.conf/make.conf is now clean after building, if a user have
some custom option in them make update-packages will not know this and will
try to build packages with some non-existant files.

Fixes:	256f4f17a120 ("jail -cu: Don't include build-time make.conf files into jail's clean snapshot")