Require ZROOTFS parent is already setup; avoid shadowing parent
If someone uses ZFOOTFS == BASEFS, then it is possible for a
zroot/usr/local/poudriere to be created with `zfs create -p` which
means a zroot/usr/local gets created. That shadows the real /usr/local
and user's curse the sky for losing data. Avoid that by requiring
the parent is already created.

Fixes #1191