noclobber: Don't force hide stderr.
Leave it up to the caller.

The use of hiding or not hiding stderr can convey expectations in the caller.
If the caller hides stderr then it may exist but are okay with not overwriting
the data. If the error is not hidden then it conveys that the value is expected
to not be set yet, that failing to set the new value is an error, so any
errors from it are valid to see.