shlibs: Don't consider base libraries for inspection.
With pkg-2.0.6+ it now registers base libraries as required in the
packages. The shared library inspection phase is because we do not know
if a dependency has a newer shared library and missed a PORTREVISION
chase. For libraries not in packages we can do this inspection before
any packages are processed. In practice if the jail is updated then
.jailversion changes and all packages are considered stale and rebuilt.

To determine if a package needs to be inspected we ignore libraries
provided for in the jail's clean snapshot. For example, if the jail
provides libc.so.7 and a package requires libc.so.7 then we trim
libc.so.7 from its required shared libraries list for the sake of
inspection. If it requires libc.so.7 and the jail provides libc.so.8
then the package will _not_ skip the inspection - as the jail's clean
snapshot is not providing for the library. It will later inspect the
package and find nothing provides <current> libc.so.7 anymore and force
a rebuild of the package to find the new libc.so.8 in the jail, or
find libc.so.7 provided for by a new package dependency, like for
pkgbase.