testport: Fix potential stats error.
If the package to be tested already exists then it is added into the
"queued"/"tobuild" lists. If it does not exist then it is not added to that
list. It is then built and added to the "built" list. This results in a stats
accounting error.  Fix this by explicitly deleting the package before
calculating the build.