sh: Drop -x patch preventing EPIPE workaround.
If set -x is set before a trap is invoked, the trap itself is traced out
to stderr. The problem comes when stderr is gone. A workaround is to
use a pattern such as:
	trap "{ set +x; } 2>/dev/null; trap_handler" SIG
This will cause the trap_handler to be called without set -x set and
avoids EPIPE.