native sh critical_section: Deal with command substitution.
In html_json it may be running $(wc -l) in the critical_section when it
receives a SIGTERM. The $() did not inherit the ignored traps so it
exits 143 and returns a blank value. This issue has been around forever.
It has also been seen in $(clock) calls.

There is still a race before the traps get inherited but there's
not much else to do, and the fallout is rare and insignificant.

With the builtin sh nothing is changed.