pkgqueue_get_next: Ensure returning empty is an empty ready-to-build queue.
Note the queue may have waiting-for-dependency items but it should
not have anything in the ready-to-build queue.

Similar to the comments in pkgqueue_empty() we need to ensure
no blank jobs are returned. It is possible for a race to happen
where pkgqueue_get_next() returns a blank job, build_queue()
gets it and checks pkgqueue_empty() and sees it is not empty,
loops around and runs the sanity check due to idle builders
and calls crashed_build() on an item that just got balanced
and was temporarily missing in the find(1) call which itself
is not an atomic/locked view of the queue.

This became apparent after making pkgqueue_balance_pool() into a
separate worker process. The first job in the queue would sometimes
get moved while the first pkgqueue_get_next() was called and
immediately get "stuck_in_queue".