Rework repositories configuration for FreeBSD 15.0
Older versions of FreeBSD had a single package repository, named
"FreeBSD".  FreeBSD 14.3 introduced a second repository for kernel
modules, named "FreeBSD-kmods".  FreeBSD 15.0 add support for managing
the base system with packages (aka pkgbase), and to avoid confusion for
users and emphasis they are related to ports and not the base system,
the mentioned repositories were respectively renamed "FreeBSD-ports" and
"FreeBSD-ports-kmods".

In order to fetch latest packages, poudriere tune the "FreeBSD" repo,
and attempts were made to ignore the "FreeBSD-kmods" repository
(#1218, #1228), unfortunately with the changes described above in
FreeBSD 15, it is broken again.

In order to avoid this issues, disable all known variations of the
default FreeBSD repositories, and setup a single "Poudriere" repository
configured to our wills.

This will allow us to have full control of that repository without being
impacted by changes on the FreeBSD side.