Always set a higher priority for repo overrides
Two configuration files for the same repository, with the same priority
(default to 0) but with a different values for `enabled` result in the
actual value of `enabled` to be dependent of the order of evaluation of
the files.  This is likely to cause problem one day.  Make sure we set
an explicit priority when overriding repositories.

While here, use conistent syntax/style for the configuration files.

(cherry picked from commit e98aa071fa116c61ef3e7c072bb7f84db4426074)

Conflicts:
	src/share/poudriere/common.sh