Avoid rounding down all values
The progress bar is composed of 6 segments, which where all rounded down
before being drawn.  Up to almost 6% of the progress bar is therefore
empty in the worst case scenario when the progress is 100%.

Skip the rounding and use the real value to compute each section size,
resulting in 100% completion being rendered as a filled-in progress-bar.