Improve consistency in status ordering
We have stats and reports by status, but the order of these status is
not consistent across the summary table on top of the page, each status
section in the body of the page and the progress-bar at the bottom of
the page.

Reorder each bit to have consistent ordering:
* queued (not shown in the progress-bar as it is the full bar)
* fetched
* ignored
* skipped
* inspected
* failed
* built
* remaining

Section in the body are listed in reverse order (as before).