dirwatch: Fix startup race when used for queue watching.
This tool is only used for watching dirs for new work items
in a context where missed items would not be good. But its
name implies it's fine to start non-empty.

If used in a context of wanting to keep the dir empty, it
is possible to block forever on the new dir waiting for new
files after a file was just written. Which could cause
the item in the queue to be lost until the next event came in.