bulk: Fix queue race
There was a race where a job could be found via the `kill -0` check and then a
new job sent to that builder, and then that first job found in the job pipe and
marked done again which crashed the new job running on the builder. Fix this by
using a unique job id.

Commit c9d494babb made this more likely, but it has been around since at
least commit 1ee7be57 which attempted to fix it but was not sufficient.

Add more tests that cover these cases:
- build_pkg() set -e
- build_pkg() SIGTERM
- build_pkg() err()

Fixes #1270