html: Avoid numeric error during SIGTERM.
When the html_json worker is killed it is possible this
`$(print $(echo | bc))` construct returns a blank value in the inner
command substitution which then results in a "expected numeric value"
error. It shows up easily if `stress_snapshot` is looped with
`HTML_JSON_UPDATE_INTERVAL=0.1` with
bulk-build-inc-nested-dep-middle-unneeded-without-recursive-delete-misordered.sh
test. With the temporary that avoids the nested command substitution the
problem does not show up.

Unlike other code in html.sh the stress_snapshot() function is not
protected by a critical section, so signals come in fine.