timestamp: Partially revert 728c011c
This change removed the `const` on prefix_stdout, prefix_stdin, and
time_start. This is not removed by the compiler, and does serve a
purpose beyond return values. Writes through the pointer, such as
`prefix_stdout[0] = 'a'` would cause a compiler error. The full
intention is that the variable is immutable once set.
`const char * const` could achieve that but is disallowed by the
compiler due to immutable pointer.

Issue #1103